Fix negative size in export
When creating a rectangle (or ellipse) while dragging from right to left, and bottom to top (i.e. creating negative height and width), the exported image is not in view. This is fixed by changing the sizes to positive values and translate the drawn object to compensate. The end result looks the same.